tests/requirements.txt

absl-py==0.7.1
aiohttp==3.6.2
alabaster==0.7.12
altair==4.1.0
argcomplete==1.10.0
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
async-timeout==3.0.1
asyncio==3.4.3
#atc-rl-env==0.1.1
atomicwrites==1.4.0
attrs==19.1.0
autopep8==1.4.4
#aviary==0.1.0
Babel==2.8.0
backcall==0.1.0
base58==2.0.0
bayesian-optimization==1.0.1
bcrypt==3.1.6
beautifulsoup4==4.8.0
bleach==3.1.4
blinker==1.4
blitz-bayesian-pytorch==0.2.3
boto==2.49.0
boto3==1.12.20
botocore==1.15.20
box2d==2.3.2
box2d-py==2.3.8
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
docopt==0.6.2
docutils==0.16
docx2txt==0.8
EasyProcess==0.2.7
EbookLib==0.17.1
entrypoints==0.3
enum-compat==0.0.3
extract-msg==0.23.1
findspark==1.3.0
flake8==3.8.2
future==0.18.2
gast==0.2.2
gensim==3.8.1
geographiclib==1.50
geojson==2.5.0
geopy==1.22.0
GPy==1.9.9
GPyOpt==1.2.5
grpcio==1.11.0
gym==0.12.1
gym-unity==0.1.1
h5py==2.10.0
html5lib==0.9999999
idna==2.9
idna-ssl==1.1.0
imagesize==1.2.0
IMAPClient==2.1.0
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
jedi==0.15.2
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
keyring==21.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
llvmlite==0.32.1
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
#mkl-fft==1.0.12
#mkl-random==1.0.2
mlagents==0.4.0
mlagents-envs==0.8.1
more-itertools==7.0.0
msgpack==1.0.0
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.6
nbstripout==0.3.7
nltk==3.4.5
notebook==5.7.8
numba==0.49.1
numpy==1.19.0rc1
numpydoc==0.9.2
olefile==0.46
opencv-python==4.1.0.25
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
paramiko==2.7.1
paramz==0.9.5
parso==0.5.2
pathtools==0.1.2
patsy==0.5.1
pbr==5.4.5
pdfminer.six==20181108
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.12.1
psutil==5.6.3
py==1.8.0
py4j==0.10.7
pyclipper==1.1.0.post3
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.7
pydeck==0.3.1
pydocstyle==4.0.1
#PyDodo==0.1.0
pyflakes==2.2.0
pygame==1.9.6
pyglet==1.3.2
Pygments==2.4.0
pylint==2.4.4
PyNaCl==1.3.0
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
#pypiwin32==223
pyproj==2.6.1.post1
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQtWebEngine==5.15.0
pyrsistent==0.15.2
PySocks==1.7.1
pyspark==2.4.5
pytest==5.2.0
python-dateutil==2.8.0
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python-pptx==0.6.18
pytz==2020.1
#PyVirtualDisplay==0.2.3
#pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.4
QtPy==1.9.0
rake-nltk==1.0.4
requests==2.23.0
rope==0.17.0
Rtree==0.8.3
s3transfer==0.3.3
scikit-learn==0.22.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.12.0
smart-open==1.9.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
SpeechRecognition==3.8.1
Sphinx==3.0.4
sphinx-autorun==1.1.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.2
spyder==4.1.3
spyder-kernels==1.9.1
statsmodels==0.11.1
streamlit==0.60.0
tensorboard==1.7.0
tensorflow==1.7.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
textract==1.6.3
toml==0.10.1
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.4
typing-extensions==3.7.4.2
tzlocal==1.5.1
ujson==1.35

urllib3==1.25.8
validators==0.15.0
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wget==3.2
